将InstanceListener与AssignmentListener从接口改为类。
