+++ libevent-api/pom.xml
@@ -12,16 +12,6 @@
     <artifactId>libevent-api</artifactId>
 
     <name>libevent-api</name>
-    <!-- FIXME change it to the project's website -->
     <url>http://www.example.com</url>
-
-    <dependencies>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>4.11</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
 
 </project>
