+++ libevent-jupiter/pom.xml
@@ -16,12 +16,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>4.11</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>com.personal.oyl</groupId>
             <artifactId>libevent-api</artifactId>
             <version>1.0.0</version>
