+++ libevent-jupiter/src/main/java/com/personal/oyl/event/jupiter/AssignmentListener.java
@@ -1,8 +1,27 @@
 package com.personal.oyl.event.jupiter;
 
+
 /**
  * @author OuYang Liang
  */
-public interface AssignmentListener {
-    void onChange(String assignment);
+public class AssignmentListener {
+
+    private EventTransportMgr manager;
+
+    public AssignmentListener(EventTransportMgr manager) {
+        this.manager = manager;
+    }
+
+    public void onChange(String assignment) {
+        if (null == assignment || assignment.trim().isEmpty()) {
+            return;
+        }
+
+        manager.stopAll();
+        String[] parts = assignment.split(JupiterConfiguration.GROUP_SEPARATOR);
+
+        for (String part : parts) {
+            manager.start(Integer.parseInt(part.trim()));
+        }
+    }
 }
