+++ libevent-kafka/pom.xml
@@ -17,12 +17,6 @@
 
     <dependencies>
         <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>4.11</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>com.personal.oyl</groupId>
             <artifactId>libevent-jupiter</artifactId>
             <version>1.0.0</version>
